refactor: extract ResponseMetaData type for LLM response metadata
